+++ .gitignore
@@ -173,6 +173,7 @@ poetry.toml
 
 # LSP config files
 pyrightconfig.json
-.opencoder
 .vscode
 # End of https://www.toptal.com/developers/gitignore/api/python
+
+.opencodr
