+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "opencodr"
-version = "0.1.6"
+version = "0.1.7"
 description = "open source terminal-based coding agent"
 readme = "README.md"
 authors = [
@@ -88,8 +88,8 @@ dev = [
 ]
 
 [project.scripts]
-opencoder = "opencoder.main:main"
-opencoder-server = "opencoder.mcp_server:main"
+opencodr = "opencodr.main:main"
+opencodr-server = "opencodr.mcp_server:main"
 
 
 [build-system]
@@ -97,7 +97,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel.force-include]
-"./src/opencoder/assets" = "assets"
+"./src/opencodr/assets" = "assets"
 
 [tool.pytest.ini_options]
 asyncio_mode = "strict"
